Handle serverside updates by manually calling the handler
This approach should be used for all the stores to guarrantee all
needed sideeffects are processed.

Additionally all handlers should check if the data recived (from update)
is the same or newer to prevent duplicates

The server still needs to subscribe to changes made by the players.
This may change if all mutation is handled by the server, which then
would recive all mutations as a response and could update itself
manually. This would simplify the serverside game with not requiring
subsciptions